# docs/swagger.yaml
basePath: /
definitions:
...
  dto.WebhookConfigRequest:
    properties:
      active:
        example: true
        type: boolean
      events:
        example:
        - message
        - receipt
        - presence
        items:
          type: string
        minItems: 1
        type: array
      url:
        example: https://my-app.com/webhook
        type: string
    required:
    - events
    - url
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para integração com WhatsApp usando whatsmeow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Zemeow WhatsApp API
  version: "1.0"
paths:
...
  /webhooks/events:
    get:
      consumes:
      - application/json
      description: Retorna lista completa de todos os eventos de webhook disponíveis
        organizados por categoria
      produces:
      - application/json
      responses:
        "200":
          description: Lista de eventos disponíveis
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Listar eventos de webhook
      tags:
      - webhooks
  /webhooks/sessions/{sessionId}/find:
    get:
      consumes:
      - application/json
      description: Retorna lista de webhooks configurados para uma sessão específica
      parameters:
      - description: ID da sessão
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de webhooks configurados
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Acesso negado
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Buscar webhooks configurados
      tags:
      - webhooks
  /webhooks/sessions/{sessionId}/set:
    post:
      consumes:
      - application/json
      description: Configura um novo webhook para uma sessão específica
      parameters:
      - description: ID da sessão
        in: path
        name: sessionId
        required: true
        type: string
      - description: Configuração do webhook
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/dto.WebhookConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Webhook configurado com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Dados inválidos
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Acesso negado
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Configurar webhook
      tags:
      - webhooks
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"